feat: :sparkles: Implementado rota e consulta do histórico patrimonial

+++ backend/core/dto/portfolio.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass
-from decimal import Decimal
 
 from backend.core.dto.base import BaseDTO
 from backend.core.dto.patrimonial_history import PatrimonialHistoryDTO
@@ -8,7 +7,7 @@ from backend.core.dto.position import PositionDTO
 
 @dataclass(frozen=True, slots=True, kw_only=True)
 class PortfolioDTO(BaseDTO):
-    cash: Decimal
+    cash: float
     variable_income: list[PositionDTO]
     # fixed_income: list[FixedIncomePositionDTO]
     patrimonial_history: list[PatrimonialHistoryDTO]
